+++ views/partials/my-posts-details.handlebars
@@ -7,7 +7,7 @@
   </div>
   <div class="flex justify-center w-full md:p-2 border-b-2 border-blue-300">
     <div id="card"
-      class="flex-row w-full p-1 rounded-sm drop-shadow-md hover:shadow-md hover:bg-slate-400 active:shadow-lg cursor-pointer hover:bg-slate-500">
+      class="flex-row w-full p-1 rounded-sm drop-shadow-md hover:shadow-md active:shadow-lg cursor-pointer hover:bg-slate-500">
       <div class="pointer-events-none">
         <h2 class="text-gray-100 text-md md:text-xl md:mb-1 font-medium"><span
             id="title">{{{post.mainpost.title}}}</span>
